add a test. Fix lots of lint warnings.

+++ jhlabs/jhlabs/source/com/jhlabs/image/EqualizeFilter.java
@@ -42,7 +42,7 @@ public class EqualizeFilter extends WholeImageFilter {
 				for (j = 1; j < 256; j++)
 					lut[i][j] = lut[i][j-1] + histogram.getFrequency(i, j);
 				for (j = 0; j < 256; j++)
-					lut[i][j] = (int)Math.round(lut[i][j]*scale);
+					lut[i][j] = Math.round(lut[i][j]*scale);
 			}
 		} else
 			lut = null;
@@ -54,7 +54,7 @@ public class EqualizeFilter extends WholeImageFilter {
 				i++;
 			}
 		lut = null;
-		
+
 		return inPixels;
 	}
